+++ spec/support/initialize_verifier.rb
@@ -1,0 +1,10 @@
+module InitializeVerifier
+  def initialize(*args)
+    @initialized = true
+    super
+  end
+
+  def initialized?
+    @initialized ||= false
+  end
+end