+++ spec/unit/active_attr/attributes_spec.rb
@@ -7,6 +7,7 @@ module ActiveAttr
 
     let :model_class do
       Class.new do
+        include InitializeVerifier
         include Attributes
         attribute :name
         attribute :amount
@@ -140,6 +141,12 @@ module ActiveAttr
           subject.name = "Ben"
           subject.attributes.should include("name" => "Benjamin")
         end
+      end
+    end
+
+    describe "#initialize" do
+      it "invokes the superclass initializer" do
+        should be_initialized
       end
     end