+++ spec/unit/active_attr/mass_assignment_spec.rb
@@ -5,6 +5,7 @@ module ActiveAttr
   describe MassAssignment, :mass_assignment do
     subject do
       model_class.class_eval do
+        include InitializeVerifier
         include MassAssignment
       end
     end