+++ spec/unit/active_attr/strict_mass_assignment_spec.rb
@@ -5,6 +5,7 @@ module ActiveAttr
   describe StrictMassAssignment, :mass_assignment do
     subject do
       model_class.class_eval do
+        include InitializeVerifier
         include StrictMassAssignment
       end
     end